Update all tests except test_HAL.py for new basis
