+++ tests/conftest.py
@@ -72,7 +72,7 @@ def fit_model_all_info(fit_data):
     fit_configs, _, E0s, data_keys, weights = fit_data
 
     basis_info = {'elements': ['Al'], 'cor_order': 2, 'maxdeg': 6,
-                  'r_cut': 3.75, 'r_in': 2.0, 'r_0': 2.0 * 1.414, 'pairs_r_dict': {}}
+                  'r_cut': 3.75, 'smoothness_prior': None}
     B_len_norm = basis.define_basis(basis_info)
 
     solver = BayesianRidge(fit_intercept=False, compute_score=True)
@@ -92,32 +92,3 @@ def fit_model_all_info(fit_data):
 @pytest.fixture(scope="module")
 def fit_model(fit_model_all_info):
     return fit_model_all_info[0]
-
-
-@pytest.fixture(scope="module")
-def fit_model_smooth_all_info(fit_data):
-    fit_configs, _, E0s, data_keys, weights = fit_data
-
-    basis_info = {'elements': ['Al'], 'cor_order': 2, 'maxdeg_ACE': 6, 'maxdeg_pair': 8,
-                  'r_cut_ACE': 3.75, 'r_cut_pair': 3.75, 'r_0': 2.0 * 1.414, 'agnesi_q': 4}
-    B_len_norm = basis.define_basis(basis_info, 'ACEHAL.bases.smooth')
-
-    solver = BayesianRidge(fit_intercept=False, compute_score=True)
-
-    n_observations = len(fit_configs) * (1 + 6) + 3 * np.sum([len(at) for at in fit_configs])
-
-    # make deterministic
-    rng = np.random.default_rng(seed=10)
-    calc, Psi, Y, coef, prop_row_inds = fit.fit(fit_configs, solver, B_len_norm, E0s,
-                                     data_keys=data_keys, weights=weights,
-                                     n_committee=8, rng=rng,
-                                     return_linear_problem=True)
-
-    return calc, Psi, Y, coef, prop_row_inds, n_observations, B_len_norm
-
-
-@pytest.fixture(scope="module")
-def fit_model_smooth(fit_model_smooth_all_info):
-    return fit_model_smooth_all_info[0]
-
-
