+++ tests/julia_assets/Project.toml
@@ -1,5 +1,6 @@
 [deps]
 ACE1 = "e3f9bc04-086e-409a-ba78-e9769fe067bb"
+ACE1pack = "8c4e8d19-0bd6-4234-8309-7210652e3178"
 ACE1x = "5cc4c08c-8782-4a30-af6d-550b302e9707"
 ASE = "51974c44-a7ed-5088-b8be-3e78c8ba416c"
 JuLIP = "945c410c-986d-556a-acb1-167a618e0462"
